fix: move locking into the utils module
