+++ src/gpu/mod.rs
@@ -37,26 +37,3 @@ use ocl::Device;
 lazy_static::lazy_static! {
     pub static ref GPU_NVIDIA_DEVICES: Vec<Device> = get_devices(GPU_NVIDIA_PLATFORM_NAME).unwrap_or_default();
 }
-
-#[derive(Debug)]
-pub struct LockedFile(File);
-
-use std::fs::File;
-
-use std::io;
-pub const LOCK_NAME: &str = "/tmp/bellman.lock";
-pub fn lock() -> io::Result<LockedFile> {
-    let file = File::create(LOCK_NAME)?;
-
-    #[cfg(feature = "gpu")]
-    {
-        use fs2::FileExt;
-        file.lock_exclusive()?;
-    }
-
-    Ok(LockedFile(file))
-}
-
-pub fn unlock(lock: LockedFile) {
-    drop(lock);
-}
