+++ src/gpu/utils.rs
@@ -1,9 +1,11 @@
 use crate::gpu::error::{GPUError, GPUResult};
 use ocl::{Device, Platform};
 
+use fs2::FileExt;
 use log::info;
 use std::collections::HashMap;
-use std::env;
+use std::fs::File;
+use std::{env, io};
 
 pub const GPU_NVIDIA_PLATFORM_NAME: &str = "NVIDIA CUDA";
 // pub const CPU_INTEL_PLATFORM_NAME: &str = "Intel(R) CPU Runtime for OpenCL(TM) Applications";
@@ -61,4 +63,21 @@ pub fn get_core_count(d: Device) -> GPUResult<usize> {
             msg: "Device unknown!".to_string(),
         }),
     }
+}
+
+#[derive(Debug)]
+pub struct LockedFile(File);
+
+pub const LOCK_NAME: &str = "/tmp/bellman.lock";
+
+pub fn lock() -> io::Result<LockedFile> {
+    let file = File::create(LOCK_NAME)?;
+
+    file.lock_exclusive()?;
+
+    Ok(LockedFile(file))
+}
+
+pub fn unlock(lock: LockedFile) {
+    drop(lock);
 }
